move array operations to isolate file
